Pending tests for ApplicationHelper#show_message
There are a couple of tricky things in terms of providing test coverage here. First, it'll have to be done in the context of a controller action, so that flash is defined. Second, we'll have to fake configuration values so that we're not relying on the git-ignored application.yml file of whatever version of this app is running the test.

I'll have to come back to this later.