Wrote ApplicationTemplate#new
Now it creates a template, a draft for that template, and redirects to
its edit page. Also changed model validations for the application
template. It was validating the uniqueness of the position, which
was dumb. I may or may not have broken the 'save changes and continue
editing' button, more on that to follow.