Changed variable, added uniqueness validation
In ApplicationTemplate#new, template did not need to be an instance
variable. In the application template model, position did in fact need
a uniqueness validation. I took it out initially because I got an error
message of "position has already been taken" but this may have been
because of something else I did. Now, the error is "number has already
been taken" referring to the question.save in create_draft in the
application template model. Tests do not pass because I have not
finished testing #new. I have opened a new branch for this, though.