Downloaded a nice sortable table
The DataTables gem. It's pretty neat. One thing that's weird is
that the little arrows that indicate you can sort do not show up
unless the image files are in the same directory as the CSS.
The css file will not 'see' them if they are in vendor/assets/images,
only if they are in vendor/assets/stylesheets. I don't know, man.