Bump Ruby version and update gems
Pessimistic version control on Rails gem.

Received deprecation warning about calling _path from a mailer, which is
good since that didn't make sense to do anyway. Changed to _url and then
needed to update the environment configs to account for this. If we
intend to send email with inward-pointing URLs in the future in other
apps, we will need to make the same change.