Only "refuse" a coverage drop in CI
If we just always refuse coverage, then it's harder to lotice that
coverage _has_, in fact, dropped since Simplecov just won't re-write the
coverage file.

This way, a dev can see the change to `.last_run.json` and either fix it
or commit it, at which point said drop can be discussed and perhaps
justified in a code-review.