Fix DataTables initial ordering. (#890)
* Order collections to match DataTables default

DataTables orders by column 0, ascending, by default. We could override
that to be unordered to start with, but it's actually not a bad default
ordering in any of the three places we use it.

* Fix tests dependent on specific controller code

These are super fragile, but lets just patch them up for now.

With the exception that I particularly don't like expectations on
`assigns`.